requirements.txt
streamlit
opencv-python
paddleocr
matplotlib
ultralytics
Pillow
numpy
